Add OpenWebUI test case generation and refactor related constants
